pal: Add palEcmat
Like slaEcmat but uses IAU 2006 precession model rather than 1980. Also,
the input date is now specified in TT instead of TDB. The difference
between TT and TDB is around one to two millisecond, or about 0.02
arc-seconds. So the cost of converting TDB to TT in every call to palEcmat
is probably not justified.