Rejig configure.ac to work outside of Starlink
Add configure tests to allow SOFA to be found outside of a Starlink
build tree. We now check for $STARLINK and if that is set we
assume that starutil and sofa are available. If it is not set we
test for SOFA and starutil and abort if SOFA can not be found.

This necessitated a change to the NOSTARUTIL define to make it
autoconf compliant and we now use HAVE_STAR_UTIL_H.