Check for Starlink on $STARLINK_DIR being set
We can not actually trust $STARLINK because the configure script
will force $STARLINK to be /star which messes the test up even
if /star does not actually exist. STARLINK_DIR is not twiddled
by configure so is a safe choice and means that we do not
need to panic if --without-starlink has been forgotten on the
configure line.